src/HotChocolate/Spatial/src/Types.Spatial/GeoJSONMultiPointInput.cs
using System.Collections.Generic;
using HotChocolate.Configuration;
using HotChocolate.Language;
using HotChocolate.Types;
using HotChocolate.Types.Descriptors.Definitions;
using NetTopologySuite.Geometries;

namespace Types.Spatial
{
    public class GeoJSONMultiPointInput : InputObjectType<MultiPoint>
    {
        private const string _typeFieldName = "type";
        private const string _coordinatesFieldName = "coordinates";
        private IInputField _typeField = default!;
        private IInputField _coordinatesField = default!;

        public GeoJSONMultiPointInput() { }

        protected override void Configure(IInputObjectTypeDescriptor<MultiPoint> descriptor)
        {
            descriptor.BindFieldsExplicitly();

            descriptor.Field(_typeFieldName).Type<EnumType<GeoJSONGeometryType>>();

            descriptor.Field(_coordinatesFieldName).Type<ListType<GeoJSONPositionScalar>>();
        }

        public override object? ParseLiteral(IValueNode literal)
        {
            if (literal is NullValueNode)
            {
                return null;
            }

            if (!(literal is ObjectValueNode obj) || obj.Fields.Count < 2)
            {
                throw new InputObjectSerializationException(
                    "Failed to serialize MultiPoint. Needs at least type and coordinate fields");
            }

            IList<Coordinate>? coordinates = null;
            GeoJSONGeometryType? type = null;

            for (var i = 0; i < obj.Fields.Count; i++)
            {
                ObjectFieldNode field = obj.Fields[i];

                switch (field.Name.Value)
                {
                    case _coordinatesFieldName:
                        coordinates =
                            (IList<Coordinate>)_coordinatesField.Type.ParseLiteral(field.Value);
                        break;
                    case _typeFieldName:
                        type = (GeoJSONGeometryType)_typeField.Type.ParseLiteral(field.Value);
                        break;
                }
            }

            if (type != GeoJSONGeometryType.MultiPoint || coordinates is null || coordinates.Count < 1)
            {
                throw new InputObjectSerializationException(
                    "Failed to serialize MultiPoint. You have to at least specify a type and" +
                    "coordinates array");
            }

            // var factory = NtsGeometryServices.Instance.CreateGeometryFactory(srid.Value);
            var points = new Point[coordinates.Count];
            for (var i = 0; i < coordinates.Count; i++) {
                points[i] = new Point(coordinates[i]);
            }

            return new MultiPoint(points);
        }

        protected override void OnAfterCompleteType(
            ICompletionContext context,
            DefinitionBase definition,
            IDictionary<string, object?> contextData)
        {
            _coordinatesField = Fields[_coordinatesFieldName];
            _typeField = Fields[_typeFieldName];
        }
    }
}
